Fix comments; move IsoMatcher tests to same package
