+++ jena-arq/src/test/java/org/apache/jena/sparql/util/TS_Util.java
@@ -29,6 +29,7 @@ import org.junit.runners.Suite.SuiteClasses ;
     TestFmtUtils.class,
     TestVersion.class,
     TestContextUtils.class,
+    TestIsoMatcher.class
 })
 public class TS_Util
 { }
